feat: Refactor AnalyzerEdit component and introduce shared components for form fields and workflow designer
- Added AnalyzerFormFields.vue for encapsulating name and description input fields with validation.
- Introduced WorkflowDesigner.vue for visual representation and management of the analysis workflow.
- Updated AnalyzerEdit.vue to utilize the new shared components, improving code organization and readability.
- Replaced inline validation schema with a dedicated analyzerValidation.ts file for better maintainability.
- Created analyzerUtils.ts for utility functions related to analyzer nodes and plugins.
- Enhanced context menu functionality for adding plugins in the workflow designer.
- Implemented logic to initialize default nodes based on available plugins.
